Extract AbstractAnnotationCompletionProposalComputer
